day3

+++ delicate-quite-4701/src/main/java/masai/com/model/Plant.java
@@ -48,7 +48,7 @@ public class Plant {
 	private Integer plantStock;
 	
 	@NotNull
-	private double plantCost;
+	private Integer plantCost;
 	
 
 }
